requirements.txt
# requirements.txt

aiobotocore
aiohappyeyeballs
aiohttp
aioitertools
aiosignal
attrs
bcrypt
blessed
blinker
boto3
botocore
Brotli
cement
certifi
cffi
charset-normalizer
click
colorama
contourpy
cryptography
cycler
defusedxml
et-xmlfile
Flask
Flask-Cors
fonttools
frozenlist
fsspec
geopandas
greenlet
gunicorn
idna
inflate64
itsdangerous
Jinja2
jmespath
kiwisolver
MarkupSafe
matplotlib
multidict
multivolumefile
numpy
odfpy
lxml
openpyxl
packaging
pandas
pathspec
pillow
psutil
psycopg
py7zr
pybcj
pycparser
pycryptodomex
pyexcel-ezodf
pyexcel-io
pyexcel-ods3
PyJWT
pyogrio
pyparsing
pyppmd
pyproj
python-dateutil
python-dotenv
pytz
PyYAML
pyzstd
requests
s3fs
s3transfer
semantic-version
setuptools
shapely
six
SQLAlchemy
termcolor
texttable
typing_extensions
tzdata
urllib3
wcwidth
Werkzeug
wrapt
XlsxWriter
yarl